Edited yet Big book and Old Book entries; new TOC
The Big Book's table of contents is now in a "yyyy-mm-dd -- title"
format, which is handier. The downside is that the table updater utility
is now out-of-date.

I've added more forwards to Old Book.

I've corrected Style/style.css so that images in a row actually get
displayed that way.