Added 'Principa Donsonia' ebook
This uses the XSLT template method of EPUB compilation I used for the
SCP books. Update Ubercoordinator and alter the directories in
principa-donsonia's Makefile if necessary.

I corrected the formatting of many of Principa Donsonia's XHTML pages,
so I copied those edited versions back into the Big Book of Key.